+++ Valigator.AspNetCore/IDescriptor.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using Valigator.Core;
+
+namespace Valigator
+{
+	public interface IDescriptor
+	{
+		DataDescriptor GetDescriptor(Type type);
+	}
+}
