+++ Valigator.AspNetCore/IValidateType.cs
@@ -1,0 +1,7 @@
+﻿namespace Valigator
+{
+	public interface IValidateType<TValue>
+	{
+		Data<TValue> GetData();
+	}
+}
