+++ Valigator.AspNetCore/ValigatorModelStateException.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
+
+namespace Valigator
+{
+	//todo: nathan
+	internal class ValigatorModelStateException : Exception
+	{
+		public ValigatorModelStateException(string name, BindingSource source, params ValidationError[] errors)
+		{
+			Name = name;
+			Source = source;
+			ValidationErrors = errors;
+		}
+
+		public string Name { get; }
+		public BindingSource Source { get; }
+		public ValidationError[] ValidationErrors { get; }
+	}
+}
